Move to a manual install rather than use ciao-install
There was an ftp error when trying to run ciao-install. I am not sure
if this is a general problem with the Docker environment that the
Binder service is using, but I have decided to try a manual installation
using Google Drive, which has the advantage of being able to install
just the segments I want, without having to worry about a fully-consistent
install. That's assuming this works.