Improve the timing section, fix some logical errors and typos
When I wrote this apparently I was getting the exact same answers
for the manual vs response approaches. I know don't so need to
point out the differences are small.

I have turned off the model caching for the time evaluation which
makes things a bit better (but actually it shouldn't really matter
as all we care about is the cost of the operations as the model
evaluation should be the same in all cases). It does make a difference
if the model evaluation cost is significantly different though (e.g.
siilar to differences in how the different approaches work).

Some documentation typos.